+++ NEWS.md
@@ -5,6 +5,9 @@
 - add/set addLandmarks(text.col = "black").
 - add Golden Square and Soho Square to addLandmarks(latlong = TRUE).
 - add point only for St Luke's Church in addLandmarks(latlong = FALSE).
+- add point and pos = 4 label for St Luke's Church in 
+  addLandmarks(latlong = TRUE)
+
 
 
 ### cholera 0.8.0